pyproject.toml
[tool.poetry]
name = "powerpoint-image-exporter"
version = "0.1.0"
description = ""
authors = ["Tom Kierzkowski <52253+tjkierzkowski@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.8"
python-pptx = "^0.6.18"

[tool.poetry.dev-dependencies]
pytest = "^6.0.1"
coverage = {extras = ["toml"], version = "^5.2.1"}
pytest-cov = "^2.10.1"
pytest-mock = "^3.3.1"
black = "^20.8b1"
flake8 = "^3.8.3"
flake8-black = "^0.2.1"
flake8-import-order = "^0.18.1"

[tool.coverage.paths]
source = ["src"]

[tool.coverage.run]
branch = true
source = ["pptx_export", "cli", "pptx_image_exporter.py"]

[tool.coverage.report]
show_missing = true
fail_under = 95

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
